refactor(android): complete joystick view decomposition
Extract the remaining right-dial segment, icon, and label rendering out of JoystickView into JoystickRightDialRenderer and remove the leftover legacy bitmap state from the view shell.

Update AGENTS, ERICK-141, and ERICK-145 to record that the Android joystick split is complete on this branch, with JoystickView reduced from 1012 lines to 490 while assembleDebug remains green.
