+++ docs/documentation/Jira/ERICK-145.md
@@ -17,16 +17,18 @@
 
 Reduce the remaining edit risk in `android/app/src/main/java/com/vatoo/erick/JoystickView.kt` by splitting rendering-heavy helpers and geometry code out of the `View` shell.
 
-Status: In progress on `ERICK-141`. `JoystickDrawingUtils.kt`, `JoystickCharacterRenderer.kt`, and `JoystickSectionRenderer.kt` have now been extracted, which reduced `JoystickView.kt` from 1012 lines to 591 lines while keeping `assembleDebug` green. The highest-risk left-dial rendering paths are now out of the `View`; remaining work is mainly the right-dial rendering loop and any final orchestration cleanup that still feels too broad.
+Status: Completed on `ERICK-141`. `JoystickDrawingUtils.kt`, `JoystickCharacterRenderer.kt`, `JoystickSectionRenderer.kt`, and `JoystickRightDialRenderer.kt` reduced `JoystickView.kt` from 1012 lines to 490 lines while keeping `assembleDebug` green. The view now keeps state, sizing, thumb handling, direction detection, and mode-specific map selection while the rendering-heavy logic lives in dedicated files.
 
 ---
 
 ## Evidence
 
-- `android/app/src/main/java/com/vatoo/erick/JoystickView.kt` started at 1012 lines and is now 591 lines after helper, character-renderer, and section-renderer extraction.
+- `android/app/src/main/java/com/vatoo/erick/JoystickView.kt` started at 1012 lines and is now 490 lines after helper and renderer extraction.
 - `android/app/src/main/java/com/vatoo/erick/JoystickCharacterRenderer.kt` now owns the left-dial character placement and text fitting flow for both 8-section and 6-section modes.
 - `android/app/src/main/java/com/vatoo/erick/JoystickSectionRenderer.kt` now owns the left-dial section fills, separator lines, and border geometry for both 8-section and 6-section modes.
-- The file still holds view state, direction detection, right-dial rendering, and mode-specific map selection in one class.
+- `android/app/src/main/java/com/vatoo/erick/JoystickRightDialRenderer.kt` now owns the right-dial segment fill, label layout, and icon rendering flow for both 8-section and 6-section modes.
+- `cd android && .\gradlew.bat assembleDebug` stayed green after each extraction slice and after the final right-dial extraction.
+- The file now mainly holds view state, direction detection, mode-specific map selection, and thumb orchestration.
 - `ERICK-141` identifies `JoystickView.kt` as one of the last Android edit surfaces that is still risky for low-context AI edits.
 
 ---
@@ -38,6 +40,8 @@ Status: In progress on `ERICK-141`. `JoystickDrawingUtils.kt`, `JoystickCharacte
 3. Preserve both 8-section and rotated 6-section drawing behavior exactly.
 4. Validate with `cd android && .\gradlew.bat assembleDebug` after each slice.
 
+Result: Completed on `ERICK-141`.
+
 ---
 
 ## Candidate Split Boundaries
@@ -45,7 +49,7 @@ Status: In progress on `ERICK-141`. `JoystickDrawingUtils.kt`, `JoystickCharacte
 1. `JoystickDrawingUtils.kt` for text sizing, icons, and color helpers
 2. `JoystickCharacterRenderer.kt` for 8-section and 6-section character placement
 3. `JoystickSectionRenderer.kt` for section/ring arc drawing and separator geometry
-4. Optional follow-on: `JoystickRightDialRenderer.kt` for the right-dial segment, icon, and label loop if `JoystickView.kt` still feels too broad
+4. `JoystickRightDialRenderer.kt` for the right-dial segment, icon, and label loop
 
 ---
 
@@ -54,3 +58,5 @@ Status: In progress on `ERICK-141`. `JoystickDrawingUtils.kt`, `JoystickCharacte
 1. `JoystickView.kt` becomes materially smaller and easier to reason about locally.
 2. Android `assembleDebug` still passes after the split.
 3. No visual or interaction regressions are introduced in 8-section or 6-section mode.
+
+Status: Met on `ERICK-141`.
